pyproject.toml
[project]
name = "compilanceqapipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.25.2",
    "azure-monitor-opentelemetry>=1.8.6",
    "azure-search-documents>=11.6.0",
    "azure-storage-blob>=12.28.0",
    "fastapi>=0.133.1",
    "firecrawl-py>=4.18.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.9",
    "langsmith>=0.7.7",
    "opentelemetry-instrumentation-fastapi>=0.60b0",
    "pypdf>=6.7.3",
    "uvicorn>=0.41.0",
    "yt-dlp>=2026.2.21",
]
